backend/requirements.txt
asgiref==3.9.1
Django==4.2.26
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.12.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.11.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
